Fix zero build
