+++ src/hotspot/share/gc/shenandoah/shenandoahGeneration.hpp
@@ -38,45 +38,6 @@ class ShenandoahReferenceProcessor;
 class ShenandoahHeap;
 class ShenandoahMode;
 
-template<ShenandoahAffiliation AFFILIATION>
-class ShenandoahIncludeRegionClosure : public ShenandoahHeapRegionClosure {
-private:
-  ShenandoahHeapRegionClosure* _closure;
-
-public:
-  explicit ShenandoahIncludeRegionClosure(ShenandoahHeapRegionClosure* closure): _closure(closure) {}
-
-  void heap_region_do(ShenandoahHeapRegion* r) override {
-    if (r->affiliation() == AFFILIATION) {
-      _closure->heap_region_do(r);
-    }
-  }
-
-  bool is_thread_safe() override {
-    return _closure->is_thread_safe();
-  }
-};
-
-template<ShenandoahAffiliation AFFILIATION>
-class ShenandoahExcludeRegionClosure : public ShenandoahHeapRegionClosure {
-private:
-  ShenandoahHeapRegionClosure* _closure;
-
-public:
-  explicit ShenandoahExcludeRegionClosure(ShenandoahHeapRegionClosure* closure): _closure(closure) {}
-
-  void heap_region_do(ShenandoahHeapRegion* r) override {
-    if (r->affiliation() != AFFILIATION) {
-      _closure->heap_region_do(r);
-    }
-  }
-
-  bool is_thread_safe() override {
-    return _closure->is_thread_safe();
-  }
-};
-
-
 class ShenandoahGeneration : public CHeapObj<mtGC>, public ShenandoahSpaceInfo {
   friend class VMStructs;
 private:
